Add agent to represent Environment, do some refactoring.
Environment agent acts like a normal agent but has its own set of modules to run, and there is only one such agent. This adds the assumption that players are based on a grid, and in fact does randomly place the players on a grid.
Also add PlayerModule and EnvironmentModule classes.
Add Janitor to run cleanup method on each module after each round.
Modify the package structure so player and environment modules are separate.